refactor(routes): Separate routes from app.js
